structure and logic of the limit entity have been changed, the logic of receiving data in various services has been redesigned. some changes and corrections

+++ src/main/java/ru/benyfox/TransactionsRestApi/dto/Limit/LimitResponseDTO.java
@@ -3,18 +3,15 @@ package ru.benyfox.TransactionsRestApi.dto.Limit;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 @Getter
 @Setter
 public class LimitResponseDTO {
-    private int accountNumber;
-
-    private long limitSum;
-
+    private Integer accountNumber;
+    private BigDecimal limitSum;
     private OffsetDateTime offsetDateTime;
-
     private String limitCurrencyShortname;
-
     private boolean limitExceeded;
 }
